requirements.txt
python-telegram-bot[job-queue]==22.0
openai==1.68.2
python-dotenv==1.0.1
